more work on pattern detection
Add a bunch of candlestick functions, cfr. the candlestick classification paper.

Add a bunch of patterns, rewriting to the style of the classification paper. Change the order of a few AND statements to slightly exploit Python's lazy evaluation. Fix a few errors.

Change main to save the detected patterns to a .parquet file instead of ever expanding the dataframe, this speeds up detection especially at the end.

Calculate some stats regarding the length of the body and such.